pmb.parse.kconfig: Raise NonBugError on invalid kernel config (MR 2279)
The intent of 5258319659a4e7edb48a9accd160bc21ff8877b7 was to not print
a stacktrace when this happened. However, I still opted to raise an
exception, so you still get a stacktrace, just with a more helpful
message. There's no need to give the user a big stacktrace here, so
raise a NonBugError instead of a RuntimeError.